+++ safe-server/src/main/java/team/ybj/service/AccountService.java
@@ -4,6 +4,6 @@ import team.ybj.pojo.YbjAccount;
 
 public interface AccountService {
 
-    YbjAccount findAccountById(Integer accountId);
+    YbjAccount findAccountById(Long accountId);
 
 }
